# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education API System",
        "description": "教育管理系统API文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.localhost",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "获取课程列表",
                "operationId": "02eeda67042dba2d22a517f26fb8eaf1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "teacher_id",
                        "in": "query",
                        "description": "教师ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "创建课程",
                "operationId": "a4d6b7e7be59571034428ffbb86aaf0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "课程名称",
                                        "type": "string"
                                    },
                                    "year_month": {
                                        "description": "年月(YYYYMM)",
                                        "type": "string"
                                    },
                                    "fee": {
                                        "description": "课程费用",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "获取课程详情",
                "operationId": "f170bd35e3106c9d9bf5807499c80b68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "课程ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "更新课程",
                "operationId": "4d00e54a4b57b3259c40894ad2d0c76b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "课程ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "课程名称",
                                        "type": "string"
                                    },
                                    "year_month": {
                                        "description": "年月(YYYYMM)",
                                        "type": "string"
                                    },
                                    "fee": {
                                        "description": "课程费用",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "删除课程",
                "operationId": "3a2888333ed2868b6102094f5348520e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "课程ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}